requirements.txt
click==8.0.3
pathspec==0.9.0
pytest==8.3.3
flake8==7.1.1
autopep8==2.3.1
mypy==1.11.2
pytest-cov==5.0.0
wheel==0.44.0
setuptools==75.1.0
azure-kusto-data==4.6.1
azure-kusto-ingest==4.6.1